Add files via upload
The project deals with the approximation of the inviscid Burgers’ equation which is a canonical non-linear flow equation that is used to model discontinuities in flow with a PINN. A custom loss function is curated specifically to the problem encompassing the PDE itself, the boundary and initial conditions. The latter act as regularization agents which imposes strict physical meaning and geometry in the function optimization. PINNs leverage the two fundamental tools of a neural network, namely function approximation and auto-differentiation. TensorFlow is used for the program and the result is tested against the exact solution and compared with the authors’. PINNs can approach both inference and system identification problems either with discrete or continuous resolutions. Here, continuous inference is used. The overall aim was to reduce the complexity of the model by introducing novel optimization techniques and different approaches. With respect to physics, the aim was to capture the discontinuity in velocity which occurs in the field.